feat: Code for Xception, image display,complete model for finetuning and augmentation
The changes in this commit are only partially tested as most require intensive computing.
This commit is also used to transfer this code to AWS.
Wrote the code for the XCeption model with several new ideas how to setup the top layers.
This needs to be tested on AWS with GPU and the results must be documented.
The overall function can now display the base images which are analyzed and also display
a sample image of the matching dog bread.

A big part is the complete InceptionV3 model with transfer learning with added top layers,
data augmentation and later fine tuning of the parameters.
This must also be tested and refactored / modified.